backend/requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
emoji==1.6.3
eventlet==0.36.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-SocketIO==5.3.6
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
nltk==3.9.1
pip==24.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.3
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
regex==2024.7.24
requests==2.32.3
rsa==4.9
simple-websocket==1.0.0
text2emotion==0.0.5
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
wsproto==1.2.0